Added some comments

+++ app/src/main/java/com/billchen/sepsislindt/Handlers/HandlerCode.java
@@ -7,23 +7,24 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 
+// code for message parsing
 @SuppressLint("UseSparseArrays")
 public enum HandlerCode {
-    HEART_RATE((int) 0x100),
-    RESPIRATION_RATE((int) 0x101),
-    SKIN_TEMPERATURE((int) 0x102),
-    POSTURE((int) 0x103),
-    PEAK_ACCLERATION((int) 0x104),
-    BREATHING_RAW((int) 0x105),
-    ECG((int) 0x106),
-    Spo2((int) 0x107),
+    HEART_RATE(0x100),
+    RESPIRATION_RATE(0x101),
+    SKIN_TEMPERATURE(0x102),
+    POSTURE(0x103),
+    PEAK_ACCLERATION(0x104),
+    BREATHING_RAW(0x105),
+    ECG(0x106),
+    Spo2(0x107),
 
-    GP_MSG_ID((int) 0x20),
-    BREATHING_MSG_ID((int) 0x21),
-    ECG_MSG_ID((int) 0x22),
-    RtoR_MSG_ID((int) 0x24),
-    ACCEL_100mg_MSG_ID((int) 0x2A),
-    SUMMARY_MSG_ID((int) 0x2B);
+    GP_MSG_ID(0x20),
+    BREATHING_MSG_ID(0x21),
+    ECG_MSG_ID(0x22),
+    RtoR_MSG_ID(0x24),
+    ACCEL_100mg_MSG_ID(0x2A),
+    SUMMARY_MSG_ID(0x2B);
 
 
     //GP_HANDLER_ID((int) 0x20);
